Replace enums with constants

+++ Cosette/Engine/Ai/Score/Evaluators/MaterialEvaluator.cs
@@ -7,7 +7,7 @@ namespace Cosette.Engine.Ai.Score.Evaluators
     {
         public static int Evaluate(BoardState board, float openingPhase, float endingPhase)
         {
-            return board.Material[(int)Color.White] - board.Material[(int)Color.Black];
+            return board.Material[Color.White] - board.Material[Color.Black];
         }
     }
 }
