+++ Cosette/Engine/Ai/Score/Evaluators/PositionEvaluator.cs
@@ -7,11 +7,11 @@ namespace Cosette.Engine.Ai.Score.Evaluators
     {
         public static int Evaluate(BoardState board, float openingPhase, float endingPhase)
         {
-            var whitePositionScore = board.Position[(int)Color.White][(int)GamePhase.Opening] * openingPhase +
-                                     board.Position[(int)Color.White][(int)GamePhase.Ending] * endingPhase;
+            var whitePositionScore = board.Position[Color.White][(int)GamePhase.Opening] * openingPhase +
+                                     board.Position[Color.White][(int)GamePhase.Ending] * endingPhase;
 
-            var blackPositionScore = board.Position[(int)Color.Black][(int)GamePhase.Opening] * openingPhase +
-                                     board.Position[(int)Color.Black][(int)GamePhase.Ending] * endingPhase;
+            var blackPositionScore = board.Position[Color.Black][(int)GamePhase.Opening] * openingPhase +
+                                     board.Position[Color.Black][(int)GamePhase.Ending] * endingPhase;
 
             return (int)whitePositionScore - (int)blackPositionScore;
         }
